jamp101: Fix ambiguous botlib abs calls
It looks like most of these are _intended_ to be fabs calls, but I
checked SoF2MP.exe and I'm pretty sure it uses int32 abs instead. It
doesn't look like that's what was originally intended, but "fixing" it
would be a breaking change.

Well, most of this is dead code anyway, but still.