[Draft] Add Com_ErrorLongjmp for module errors
We need to use longjumps for module errors. I thought that could be done
by exposing a different Com_Error variant to modules, but I forgot to
account for Com_Error calls within other trap calls. Having separate
versions of those for module vs. engine is not feasible, so we'll need
to use a different approach.