[Draft] Translate between exceptions and longjumps at module boundary
Throwing exceptions across module boundaries causes crashes, so we want
to use longjmp instead. But longjmp does not call destructors, so we
want to stick to exceptions within the engine, and probably also within
modules.

To that end, we want to translate between the two at the engine-module-
boundary.

This incomplete sketch outlines exception-to-longjmp translation for
trap calls. The corresponding longjmp-to-exception translation for
vmCalls is not yet implemented.

I'm using an std::list for the std::jmp_buf stack to ensure stable
addresses, but in practice I think those buffers have value semantics,
so they should be safe to copy. Subsequently, an std::vector might be a
more efficient choice.