[WIP] use templates for typesafe syscalls
I thought I could use templates to correctly distinguish the different types
that will be passed to syscalls from qvm vs. native code, but the problem is
that I still need to present these parameters as void*/void** to the module. I
can't get the TranslateG2Args helper that is supposed to automate the
conversion to work.

With the benefit of hindsight, I think changing the arguments was the wrong
approach. I should instead conditionally cast to different types inside of
those functions. I'll give that a try in another branch.