[WIP] reimplement entity mapper using references
I thought we could avoid replacing all the `.` with `->` by using references,
but in doing so I realized that this only works for bitwise identical types,
while all pointers need to go through `VM_ArgPtr`, so if we make it look the
same but behave differently, that just invited mistakes.

We should instead try a smart reference approach, something akin to
`std::reference_wrapper`, but with added qvm translation.