Use int32_t in sharedEntity_t
sharedEntity_t and everything therein is shared between engine and modules. We
must therefore ensure a stable ABI. We currently implicitly assume that `int`
is 32 bit large, let's be explicit about it instead.

If someone is using OJK on an obscure platform with non-32-bit ints, this will
be a breaking change and modules need to be recompiled. But this
standardization is necessary for QVM, where types must have a fixed layout.