[WIP] Templatize sharedEntity_t for QVM/Native
The engine needs to distinguish between the memory layouts of native and
QVM entities, but we don't want to duplicate the definition. This commit
attempts to solve the issue using templates, the header in question is
shared with the modules, which use C instead of C++. This necessitates
the use of macros.

I'm abandoning this approach for now because I've been informed that
sharedEntity_t is really mostly opaque to the modules, which use
gentity_t instead. Based on that, I'll attempt to move the definition
entirely into the engine, so it can use C++ without C backwards
compatibility macro hacks.