Move sharedEntity_t into the server
To the modules, sharedEntity_t is just an opaque pointer to the header
of a gentity_t. Only the engine needs to know how to actually interpret
that, and that interpretation actually depends on whether a native or a
QVM module is used.

Moving the definition entirely within the C++ part of the code base will
allow us to use templates to model the differences between native and
QVM without needing workarounds for the C modules.