[WIP] Distinguish native and QVM with templates
Native and QVM sharedEntity_t structs have different memory layouts
because they contain pointers. In the QVM context, these pointers need
to be resolved within the VM address space.

The previous approach was to introduce an additional level of
indirection, a sharedEntityMapper_t that was filled with pointers to the
correct field addresses in SV_LocateGameData. I dislike the memory
overhead of having a Mapper per MAX_GENTITIES, the CPU overhead of
precaching all the mappers in SV_LocateGameData, and how this approach
won't work for Vehicle_t, because the module can choose their limit.

So I'm exploring an alternative approach: Using C++'s powerful type
system to model the differences between QVM and Native.
sharedEntityMapper_t is now a template, with QVM and Native versions
sharedEntity_qvm_t and sharedEntity_native_t that correctly model the
corresponding memory layouts. In the QVM type, pointers are modelled
using a custom pointer wrapper that automatically performs the QVM
address space translation.

This way, we can model the differences without memory or cpu overhead
(beyond the unavoidable overhead of resolving QVM addresses).

The downside of this is that any function that wants to use
sharedEntityMapper_t needs to either be a template function, or needs
separate branches for QVM and Native. The impact of that is massive.

I have fully adjusted the Icarus functions so that GameInterface.cpp and
Q3_Interface.cpp can be compiled as a proof of concept, but a lot of
other module calls are still missing.

Side note: I've changed the gEnts argument to SV_LocateGameData from a
sharedEntity_t pointer to a void pointer. On the ABI level, this should
be a backwards compatible change, as it's just a pointer either way.
While it is correct that the passed pointer points to a sharedEntity_t,
typing it thusly incorrectly suggests that it points to an array of
sharedEntity_t, when it actually points to a gentity_t array of variable
element size. I therefore find it more honest to treat it as an opaque
pointer.