Use union for pointers in sharedEntityMapper_t
It seems dangerous to pretend that the pointers to pointers inside
sharedEntityMapper_t are all ordinary pointers. Instead, let's use a
union type to clearly distinguish between ordinary native pointers and
32-bit pointers into QVM memory that need to be translated first.

I chose to only use a single union-pointer-type, instead of one per
pointed-to type, for simplicity. I added some light type annotations
using no-op defines, but there's no actual type checking.

I did all this in the pursuit of properly marking Vehicle_t as having
different memory layouts in native and QVM. It is only used for
null-checks at the moment, which is safe, but if somebody were to try to
access its fields, they need to be aware of the type mismatch.